requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
certifi==2022.9.14
cffi==1.15.1
channels==4.0.0
channels-auth-token-middlewares==1.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.1
daphne==4.0.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.5
Django==4.0.7
django-allauth==0.51.0
django-appconf==1.0.5
django-cleanup==6.0.0
django-cors-headers==3.10.1
django-model-utils==4.2.0
django-versatileimagefield==2.2
djangorestframework==3.14.0
drf-spectacular==0.24.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jsonschema==4.16.0
msgpack==1.0.4
oauthlib==3.2.1
packaging==21.3
Pillow==9.2.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-magic==0.4.27
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.10.0
txaio==22.2.1
types-cryptography==3.3.23
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
zope.interface==5.5.1
